fix da build after makin da ch ch changes. This:
  - abstract away interfaces for node.go.
    To be thought about / harnessed in near future.
  - init bft committee: re-align project structure
  - add init package in manifest
