+++ config/logistics_bft.cfg
@@ -1,0 +1,20 @@
+[general]
+# overall config
+pem_keystore_path = /var/lib/truechain/keys
+# config_folder =
+
+#########################################
+# node specific
+[node]
+# peer_network_file location is currently defined in local_config.py
+ledger_location = /var/log/truechain
+
+#########################################
+# logging specific
+[log]
+root_folder = /var/log/truechain
+server_logfile = engine.log
+client_logfile = client.log
+
+# 1024*1024*4
+max_log_size = 4194304
