--- pbft/src/pbft-core/pbft-utils.go
@@ -1,155 +1,0 @@
-/*
-Copyright (c) 2018 TrueChain Foundation
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package pbft
-
-import (
-	"bytes"
-	"fmt"
-	"math/big"
-	"math/rand"
-	"path"
-	"reflect"
-
-	pb "pbft-core/fastchain"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/trie"
-	"github.com/golang/protobuf/proto"
-
-	ethcrypto "github.com/ethereum/go-ethereum/crypto"
-)
-
-// HashLength defines length of hashes (32 bytes)
-const HashLength = 32
-
-var hashT = reflect.TypeOf(Hash{})
-
-// Hash defines a hash of 32 bytes
-type Hash [32]byte
-
-// BytesToHash sets b to hash.
-// If b is larger than len(h), b will be cropped from the left.
-func BytesToHash(b []byte) Hash {
-	var h Hash
-	h.SetBytes(b)
-	return h
-}
-
-// BigToHash sets byte representation of b to hash.
-// If b is larger than len(h), b will be cropped from the left.
-func BigToHash(b *big.Int) Hash { return BytesToHash(b.Bytes()) }
-
-// HexToHash sets byte representation of s to hash.
-// If b is larger than len(h), b will be cropped from the left.
-func HexToHash(s string) Hash { return BytesToHash(common.FromHex(s)) }
-
-// Bytes gets the byte representation of the underlying hash.
-func (h Hash) Bytes() []byte { return h[:] }
-
-// Big converts a hash to a big integer.
-func (h Hash) Big() *big.Int { return new(big.Int).SetBytes(h[:]) }
-
-// Hex converts a hash to a hex string.
-func (h Hash) Hex() string { return hexutil.Encode(h[:]) }
-
-// TerminalString implements log.TerminalStringer, formatting a string for console
-// output during logging.
-func (h Hash) TerminalString() string {
-	return fmt.Sprintf("%x…%x", h[:3], h[29:])
-}
-
-// String implements the stringer interface and is used also by the logger when
-// doing full logging into a file.
-func (h Hash) String() string {
-	return h.Hex()
-}
-
-// Format implements fmt.Formatter, forcing the byte slice to be formatted as is,
-// without going through the stringer interface used for logging.
-func (h Hash) Format(s fmt.State, c rune) {
-	fmt.Fprintf(s, "%"+string(c), h[:])
-}
-
-// UnmarshalText parses a hash in hex syntax.
-func (h *Hash) UnmarshalText(input []byte) error {
-	return hexutil.UnmarshalFixedText("Hash", input, h[:])
-}
-
-// UnmarshalJSON parses a hash in hex syntax.
-func (h *Hash) UnmarshalJSON(input []byte) error {
-	return hexutil.UnmarshalFixedJSON(hashT, input, h[:])
-}
-
-// MarshalText returns the hex representation of h.
-func (h Hash) MarshalText() ([]byte, error) {
-	return hexutil.Bytes(h[:]).MarshalText()
-}
-
-// SetBytes sets the hash to the value of b.
-// If b is larger than len(h), b will be cropped from the left.
-func (h *Hash) SetBytes(b []byte) {
-	if len(b) > len(h) {
-		b = b[len(b)-HashLength:]
-	}
-
-	copy(h[HashLength-len(b):], b)
-}
-
-// Generate implements testing/quick.Generator.
-func (h Hash) Generate(rand *rand.Rand, size int) reflect.Value {
-	m := rand.Intn(len(h))
-	for i := len(h) - 1; i > m; i-- {
-		h[i] = byte(rand.Uint32())
-	}
-	return reflect.ValueOf(h)
-}
-
-// HashTxns returns a hash of all the transactions using Merkle Patricia tries
-func HashTxns(txns []*pb.Transaction) []byte {
-	trie := new(trie.Trie)
-	for i, txn := range txns {
-		val, _ := proto.Marshal(txn)
-		trie.Update(proto.EncodeVarint(uint64(i)), val)
-	}
-
-	return trie.Hash().Bytes()
-}
-
-// HashBlockHeader returns a hash for a block header
-func HashBlockHeader(header *pb.PbftBlockHeader) []byte {
-	headerData, _ := proto.Marshal(header)
-	return ethcrypto.Keccak256(headerData)
-}
-
-// VerifySender verifies transaction sender by matching public key obtained from transaction signature with expected public key
-// TODO This should match sender addresses later
-func VerifySender(tx *pb.Transaction, n int) ([]byte, bool) {
-	sig := tx.Data.Signature
-	txPubkey, _ := ethcrypto.Ecrecover(tx.Data.Hash, sig)
-
-	cfg := GetPbftConfig()
-	pubKeyFile := fmt.Sprintf("sign%v.pub", n)
-	fmt.Println("fetching file: ", pubKeyFile)
-	pubKey, _ := FetchPublicKeyBytes(path.Join(cfg.KD, pubKeyFile))
-
-	if bytes.Equal(txPubkey, pubKey) {
-		return pubKey, true
-	}
-
-	return nil, false
-}
