--- pbft/tests/README.md
@@ -1,23 +1,0 @@
-WIP:
-
-..add hooks for code coverage, smoke tests, SVT integrations, CI & build status.
-
-## Tests for ...
-
-### PBFT core consensus
-
-- [ ] Unit
-- [ ] Functional
-- [ ] Integration
-
-### BFT Committee
-
-### Hybrid Consensus
-
-- [ ] Smoke
-- [ ] Integration
-
-### SnailChain
-
-- [ ] Unit
-- [ ] Functional
