+++ support/scripts/check.sh
