+++ support/scripts/test.sh
